:truck: kafka/outbox package 수정 - 피드백 반영
