+++ .gitignore
@@ -12,6 +12,7 @@ build/
 *.iws
 *.iml
 *.ipr
+
 out/
 !**/src/main/**/out/
 !**/src/test/**/out/
@@ -40,3 +41,5 @@ bin/
 
 ### Mac OS ###
 .DS_Store
+
+*.xml
