+++ src/main/java/com/hhplus/concertReserv/domain/reservation/message/entity/ReservationOutbox.java
@@ -1,4 +1,4 @@
-package com.hhplus.concertReserv.domain.reservation.kafka.outbox.entity;
+package com.hhplus.concertReserv.domain.reservation.message.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -8,12 +8,11 @@ import org.hibernate.annotations.ColumnDefault;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-
 @Entity
 @Getter
 @Setter
-@Table(name= "PAYMENT_OUTBOX" ,indexes = @Index(name = "idx_payment_id", columnList = "PAYMENT_ID"))
-public class PaymentOutbox {
+@Table(name= "RESERVATION_OUTBOX" ,indexes = @Index(name = "idx_seat_user_id", columnList = "SEAT_ID,USER_ID"))
+public class ReservationOutbox {
 
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
@@ -23,11 +22,11 @@ public class PaymentOutbox {
     private String message;
     private String status;
 
-    private UUID paymentId;
-    private LocalDateTime eventCreateDt;
-    private LocalDateTime eventCompletedDt;
-
     @ColumnDefault("0")
     private Long retry;
 
+    private UUID seatId;
+    private UUID userId;
+    private LocalDateTime eventCreateDt;
+    private LocalDateTime eventCompletedDt;
 }
