+++ src/main/java/com/hhplus/concertReserv/infrastructure/kafka/reservation/ReservationMessageProducer.java
@@ -1,7 +1,7 @@
 package com.hhplus.concertReserv.infrastructure.kafka.reservation;
 
 import com.hhplus.concertReserv.domain.reservation.event.ReservationEvent;
-import com.hhplus.concertReserv.domain.reservation.kafka.ReservationMessagePublisher;
+import com.hhplus.concertReserv.domain.reservation.message.ReservationMessagePublisher;
 import com.hhplus.concertReserv.infrastructure.kafka.KafkaMessage;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -9,7 +9,6 @@ import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDateTime;
-import java.util.UUID;
 
 @Component
 @Slf4j
@@ -27,10 +26,10 @@ public class ReservationMessageProducer implements ReservationMessagePublisher {
     public void publish(ReservationEvent reservationEvent) {
         KafkaMessage<ReservationEvent> message = new KafkaMessage<>();
         message.setPublishDt(LocalDateTime.now());
-        message.setEventKey(UUID.randomUUID().toString());
         message.setPayload(reservationEvent);
 
         log.info(">>> Kafka producer Send start : Reservation Info {}",message);
+        // TODO key
         kafkaTemplate.send(reservationTopic, message);
         log.info(">>> Kafka producer Send end : Reservation Info {}",message);
     }
