+++ src/main/java/com/hhplus/concertReserv/infrastructure/spring/reservation/outbox/ReservationOutboxJPARepository.java
@@ -1,6 +1,6 @@
 package com.hhplus.concertReserv.infrastructure.spring.reservation.outbox;
 
-import com.hhplus.concertReserv.domain.reservation.kafka.outbox.entity.ReservationOutbox;
+import com.hhplus.concertReserv.domain.reservation.message.entity.ReservationOutbox;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
